Added some sounds

+++ Assets/Scripts/Lock System/KeyLocker.cs
@@ -8,7 +8,7 @@ namespace Lock_System
             NetworkVariableWritePermission.Owner);
 
         public override void Init(int userId)
-            => _registratedKey.Value = (ushort)userId;
+            => _registratedKey.Value = userId;
 
         public override bool CanBeOpened(int value)
             => _registratedKey.Value == value;
